[AArch64] Refactor section_group[] representation.
Change the behaviour of section_group[] such that .stub_sec points to
the stub section attached to the indexed section rather than the stub
section attached to the link_section pointed to be the index section.

This provides a mechanism to get to the stub section following any
input section.  While still allowing the section grouping mechanism to
find the section group stub section associated with an input section
by first following the link_sec pointer.

Change-Id: I19bce3feea77bbc71c04f2d8b3a204c527a21d9a
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>